+++ tests/dynamical/test_dynamic_interventions.py
@@ -8,13 +8,9 @@ from chirho.counterfactual.handlers import (
     MultiWorldCounterfactual,
     TwinWorldCounterfactual,
 )
-from chirho.dynamical.handlers import (
-    DynamicIntervention,
-    ODEDynamics,
-    SimulatorEventLoop,
-    simulate,
-)
-from chirho.dynamical.ops import State
+from chirho.dynamical.handlers import DynamicIntervention, SimulatorEventLoop
+from chirho.dynamical.ops import State, simulate
+from chirho.dynamical.ops.ODE import ODEDynamics
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 
 from .dynamical_fixtures import UnifiedFixtureDynamics
