+++ tests/dynamical/test_noop_interruptions.py
@@ -8,9 +8,8 @@ from chirho.dynamical.handlers import (
     PointInterruption,
     PointIntervention,
     SimulatorEventLoop,
-    simulate,
 )
-from chirho.dynamical.ops import State
+from chirho.dynamical.ops import State, simulate
 
 from .dynamical_fixtures import UnifiedFixtureDynamics, check_trajectories_match
